requirements.txt
# Core dependencies
transformers
sentence-transformers
#torch
faiss-cpu
accelerate
psutil

# Document processing
PyPDF2
python-docx
pdfplumber

# Interface
streamlit

# Utilities
numpy
pandas
tqdm
python-dotenv

# Testing
pytest
pytest-cov